Update linter.py
Auxiliary directory kill the hierarchy folder search. Here, the plugin will only look for .htmlhintrc in '~' or the directory of the inspected file, missing a .htmlhintrc file at the root of the project.
Removing the auxiliary directory will still find a .htmlhintrc file in '~' if no other config file are found while moving up from the directory of the inspected file.